+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ----
 
+# Development
+
+### 1.x.2.x  
+(24.01.2019)  
+[Based on OpenCAD 0.2.6]
+
+* Base: This is a major base upgrade.
+  * Limit number of Civ Identities
+
+----
+
 # 1.3
 
 ### 1.3.1.1  
@@ -23,7 +34,7 @@
 ATVG-CAD originaly was forked from OpenCAD 0.2.2.7 and was fastly upgraded to OpenCAD 0.2.3 and got some patches for 0.2.4. This is the update to OpenCAD 0.2.6 Development.
 
 ### 1.3.0.0
-(02.01.2019)
+(02.01.2019)  
 [Based on OpenCAD 0.2.3.2]
 
 * Mod: Api Security
@@ -31,7 +42,7 @@ ATVG-CAD originaly was forked from OpenCAD 0.2.2.7 and was fastly upgraded to Op
 # 1.2
 
 ### 1.2.0.1
-(02.01.2019)
+(02.01.2019)  
 [Based on OpenCAD 0.2.3.2]
 
 * Add: jsimport for simpler JS imports